+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.1.0] - 2024-06-12
+
+- Added English translation to the README
+- Added a function that open the directory in Code
+
 ## [1.0.1] - 2024-06-11
 
 - Added icon